registration form

+++ studybud/base/views.py
@@ -5,6 +5,9 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 
 # Create your views here.
 # rooms = [
@@ -14,12 +17,19 @@ from django.contrib.auth import authenticate, login, logout
 # ]
 
 
+######## LEFT AT 3:00
+
+
 
 def loginPage(request):
     
+    page = 'login'
+    
+    if request.user.is_authenticated:
+        return redirect('home')
     
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         
         try:
@@ -35,7 +45,7 @@ def loginPage(request):
         else:
             messages.error(request, 'Username or Password is incorrect')
     
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
     
 
@@ -43,6 +53,23 @@ def loginPage(request):
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+
+def registerUser(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request,"An error occured during registeration")
+    
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
         
@@ -68,6 +95,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     
@@ -82,10 +110,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse("Not allowed")
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -96,8 +127,12 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom(request, pk):
+    
+    if request.user != room.host:
+        return HttpResponse("Not allowed")
+    
     room = Room.objects.get(id=pk)
     if request.method == 'POST':
         room.delete()
